change TensorRank type to size_t

Former-commit-id: fdf90e3000778615ce32b7d027a1b22c0cf52a72
Former-commit-id: 882cee80e90c9774aff5bbd3287f1cd1886fb1f3
Former-commit-id: d57cc66df39d5ff46135c2774a33e4b1b66a91cd
Former-commit-id: 2d275a6f2fdbbb1e9f23bcd4a4b714273e739f02 [formerly e3c9ba5bc3f03ab4dd8a2fa7e79431bdeecd4be3]
Former-commit-id: d0e182d3336d02c6c4c57a7e42f758dc07415a39

+++ tammx/types.h
@@ -15,7 +15,7 @@ namespace tammx {
 
 struct BlockDimSpace;
 using BlockIndex = StrongNum<BlockDimSpace, int64_t>;
-using TensorRank = int;
+using TensorRank = size_t;
 struct IrrepSpace;
 using Irrep = StrongNum<IrrepSpace, int>;
 struct SpinSpace;
